#5509 IPv6 fields and network byte order support
<==== Mercurial data ====>
Branch: feature_5509_pcp

+++ common/src/utils/network/system_socket.h
@@ -542,7 +542,7 @@ private:
         void* buffer,
         unsigned int bufferLen,
         HostAddress* const sourceAddress,
-        int* const sourcePort );
+        quint16* const sourcePort );
 };
 
 #endif
